CP-1132 add getFiles to device-service
