+++ packages/app/src/files-manager/actions/get-files.action.test.ts
@@ -9,7 +9,7 @@ import { AnyAction } from "@reduxjs/toolkit"
 import { testError } from "Renderer/store/constants"
 import { getFilesRequest } from "App/files-manager/requests/get-files.request"
 import { initialState } from "App/files-manager/reducers"
-import { McUsbFile, McUsbFileType } from "App/mc-usb"
+import { McUsbFile, McUsbFileType } from "@mudita/pure"
 import { FilesManagerEvent } from "App/files-manager/constants"
 import { getFiles } from "App/files-manager/actions"
 import { GetFilesError } from "App/files-manager/errors"
