+++ packages/app/src/files-manager/actions/get-files.action.ts
@@ -8,7 +8,7 @@ import { FilesManagerEvent } from "App/files-manager/constants"
 import { GetFilesError } from "App/files-manager/errors"
 import { setFiles } from "App/files-manager/actions/base.action"
 import { getFilesRequest } from "App/files-manager/requests/get-files.request"
-import { McUsbFile } from "App/mc-usb"
+import { McUsbFile } from "@mudita/pure"
 
 export const getFiles = createAsyncThunk<Error | McUsbFile[]>(
   FilesManagerEvent.GetFiles,
